Use `parallel` package for parallel processing
The internal functions used for parallel processing of the MCMC simulations was
changed from the `foreach` package to the `parLapply` function in the `parallel`
package. This change removes the option to use the `multicore` "MC" option for
parallel processing.

Using the `parallel` package, which is a standard R package, limits the
`jagtools` dependence on 3rd party packages and simplifies the logic of the
package code.